src/model/requirements.txt
torch
lightning
wandb
transformers
pillow
numpy
einops
python-dotenv
opencv-python-headless
torchvision
accelerate